Split errors and exit error wrappers to separate files
